+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.2] - DEV
+- Add a new submenu "Operations".
+- Add horizontal and vertical mirror features.
 - Ctrl+W can be used to close a tab.
 - Various optimizations.